try to address duplicity sonar report, extract sections into components (#985)
* try to address duplicity sonar report, extract sections into components

* move checkbox options to constants file

* use formatOptions function to map options

---------
