+++ client/src/components/modal-forms/ArchiveHiredParticipantForm.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Box } from '@material-ui/core';
 import Grid from '@material-ui/core/Grid';
 import dayjs from 'dayjs';
-import { addYearToDate, getTodayDate } from '../../utils';
+import { addYearToDate, formatOptions, getTodayDate } from '../../utils';
 import { Button } from '../generic';
 import { Field, Formik, Form as FormikForm } from 'formik';
 import { RenderSelectField, RenderDateField, RenderCheckbox, RenderRadioGroup } from '../fields';
@@ -19,15 +19,6 @@ import {
   ROSUnderwayStatus,
   ROSCompleteStatus,
 } from '../../constants';
-
-/**
- * Formats for RenderSelectField
- * @param {[string]} optionList
- * @returns {[{value: string, label: string},...]}
- */
-const formatOptions = (optionList) => {
-  return optionList.map((option) => ({ value: option, label: option }));
-};
 
 const reasonOptions = formatOptions(archiveReasonOptions);
