+++ client/src/utils/formatOptions.js
@@ -1,0 +1,8 @@
+/**
+ * Formats for RenderSelectField
+ * @param {[string]} optionList
+ * @returns {[{value: string, label: string},...]}
+ */
+export const formatOptions = (optionList) => {
+  return optionList.map((option) => ({ value: option, label: option }));
+};
