+++ client/src/utils/index.js
@@ -8,3 +8,4 @@ export * from './prettifyStatus';
 export * from './gen-util';
 export * from './csvReports';
 export * from './keycloak-util';
+export * from './formatOptions';
